Move TABLE SAMPLE test to TPCH distributed queries
Table sample is handled by Presto internally by sampling splits, and for this
test to work the connector must produce multiple splits.  The TPCH connector
always produces multiple splits, so testing with just this connector is enough
to test the TABLE SAMPLE system.
